server/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pyyaml>=6.0
